Splits TikaUtil into smaller parts, separating functionality
