+++ stages/utils/tika/pom.xml
@@ -14,15 +14,9 @@
 
     <dependencies>
         <dependency>
-            <groupId>com.findwise.hydra</groupId>
-            <artifactId>hydra-api</artifactId>
-            <version>${project.parent.version}</version>
-            <type>jar</type>
-        </dependency>
-        <dependency>
             <groupId>org.apache.tika</groupId>
             <artifactId>tika-parsers</artifactId>
-            <version>1.1</version>
+            <version>1.5</version>
         </dependency>
         <dependency>
             <groupId>junit</groupId>
